services/codeact_cardcheck/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codeact-cardcheck"
version = "0.1.0"
description = "CodeAct-style agent using ast-grep to verify model cards against code"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2",
    "pyyaml",
    "nbconvert>=7.0",
    "nbclient>=0.9",
    "papermill>=2.5",
    "black>=23.0",
    "ruff>=0.1",
    "jsonschema>=4.0",
    "markdown>=3.4",
    "beautifulsoup4>=4.12",
    "python-docx>=1.0",
    "fastapi>=0.100",
    "uvicorn>=0.23",
]

[tool.setuptools]
packages = ["tools", "reporters"]

[tool.setuptools.package-data]
"*" = ["rules/*.yaml", "schemas/*.json", "sgconfig.yml"]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[tool.uv]
index-url = "https://pypi.org/simple"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

